In the Home frame, after selecting the desired items and respective quantities and proceeding to invoice window, i want to add an option of edit the quantities entered in the table. The problem is i cant set the only quantity column as editable. Tried modifying table model in the source, that also didnt worked. suggest some solution to this.
second i want to do changes to different payment frames including UI. Suggest some changes in the procedure.

in addition to this, after confirming the payment in the respective frame, i hv set a timer of some 10-15 sec. it just tells to wait for that time and then a confirmation message will appear. i want to create a frame operated and controlled by the restaurant end. and in the timer the message which was only appearing after the entered time, i want that message to appear only when some confirmation act is sensed from that restaurant end.

any other features that i have missed are welcome. I discontinued this project a long ago when i was in school, but now want to take to some kind of professional level.(i have started my first year classes 2 weeks ago only in cse).

Please ignore any format errors as this is my first upload on github. also if this project seems naive, reason being i started while i was in school. 
If this project is not worth giving time, i would be grateful to know that.

Thanks.